refact: make new file .. make_env_data.c
