+++ .gitignore
@@ -1,3 +1,4 @@
 *.o
 *.out
 *.DS_Store
+minirt
